parsers units record superscript zero for the record
broken in racket, need to know what is going on here since I copied
and pasted from the racket file yet windows has no issue reading this
python source with that char, but somehow file.read fails when it is
the racket file? is there some weird file encoding nonsense going on?